chore: Migrate repository to anelson-labs
Due to stupid GitHub limitations, there are features (specifically ARM
runners) available to an org account that simply cannot be obtained by
an individual account for any price.  For that reason I've maintained an
org account `anelson-labs`, to which I move repos once they need those
runners.

That came up very quickly with this project as I want to produce ARM
binaries, so I'm moving this repo just a few hours after it was created.