ci: Fix release-plz workflow issues
Previously I thought using GITHUB_TOKEN was a mistake here, but actually
it's a must; if I use the `RELEASE_PLZ_TOKEN` secret, which is my
personal token, the the PR is opened by me and that means I can't
approve it.  That's just sloppy.

Whilst at it, I also updated the check for the org `anelson` to
`anelson-labs` since that's where this repo lives now.