ci: Try once again to fix release-plz workflow
I think it's right now.  `release` uses the `RELEASE_PLZ_TOKEN` secret,
mapping to my personal user name, to make the tag that will trigger
`cargo dist` builds, while the `pr` job uses GITHUB_TOKEN so that the PR
that it creates is owned by the gha bot and not me, so that I can
approve the PR.

Phew.