feat: Automatically detect Github and Gitlab URLs in --git argument
When the user specifies a GitHub or Gitlab repository URL with the `--git` argument,
we can automatically recognize it and treat it as a reference to a crate in that forge.
This allows us to leverage the forge's API to look for pre-built binary releases,
which can significantly speed up installation.