feat: Implement running binaries after they are built.
I based this implementation on what `cargo run` does, which varies
depending on platform.

I don't have any automated tests of this yet, as on Unix it uses
`execve` which would definitely interfere with the test runner process.

I'll add more thorough tests that invoke the actual binary in various
situations, but that will be done later.