refactor: Enable several optional clippy lints and clean up code accordingly
I enabled the lints that Embark Studios use (see
<https://github.com/EmbarkStudios/rust-ecosystem/blob/main/lints.toml>),
and fixed the resulting warnings in the codebase.  I like all of these
additional lints, they are not onerous or pedantic, and they catch
issues that help make the code cleaner and more disciplined.

This is all the more valuable when one uses coding agents to perform
tedious tasks, as they do a lot better when they have strict guardrails
limiting their ability to run amok.