fix: Fix failing SBOM snapshot tests on Windows and Mac
When I made these snapshots I didn't think about the fact that a given
crate's SBOM will vary by platform, since it's quite common to have
target-specific dependencies.  To be extra thorough I should probably
make separate snapshot tests for each platform, but I don't think that's
worthwhile.  The point is to catch regressions in SBOM generation, and
the Linux snapshots are sufficient for that purpose.