fix: Update to tame-index 0.23.1 to fix Windows lock timeout issue
Before I could even open a PR with my fix for
EmbarkStudios/tame-index#94, @Jake-Shadle had already landed
EmbarkStudios/tame-index#95 and made a new release 0.23.1.  And on a
weekend no less.  Very impressive.

This updates to that fixed version of `tame-index`, and for good measure
adds a stress test that causes heavy lock contention on the cargo
package cache to ensure the fix is effective and possibly detect
regressions or other lock-related issues.