feat: Add new options `--no-exec` and `--list-targets` to CLI
`--no-exec` will be useful for testing, but also for scripting use cases where
you want to get the path to the built binary without executing it.
Maybe because you want to pre-heat the cache to ensure a particular
executable is available offline, or maybe you want to invoke the
executable yourself in some other way.

`--list-targets` allows users to see what binary and example targets are
available in a crate, which is useful for crates with multiple binaries or
where the user is unsure what binaries are provided.  I was not able to
find an example of a popular Rust crate that has multiple binaries, but
as my test data show it absolutely can happen and `cgx` needs to be able
to execute all of the bins in such a crate.