feat: Instrument code with `tracing` for improved logging and diagnostics
This is definitely not the end of the logging story.  The code isn't well-instrumented, the log events are not carefully designed to be user-friendly and at the right level, and there's no support for logging to a file yet.

However, this is a good start, and tests have been updated to initialize logging so that debug output is available when running tests.