feat: Add support for config files read from various locations (#18)
This is a pretty substantial new feature, without which `cgx` isn't that
useful.

Many options can now be set in a `cgx.toml`, which somewhat like
`.cargo/config.toml` can be placed in the working directory or anywhere
up the filesystem hierarchy, at the user level, and at the system level.

To implement this I added some config files into `testdata/` and added a
lot of tests to exercise the construction of `CrateRef` and
`BuildOptions` from the CLI args, which now must also consider the
`Config` object.

This is definitely not the final form of config, but it's enough to be
able to pin tools to specific versions which is what I need for my own
use case in my Rust projects.