ci: Add special-case handling for dependabot PRs in CI workflow
Dependabot is a very prolific opener of PRs, by design.  As much as I
would like to run the full CI suite on every dependency change, I pay
for all of those GHA minutes out of my own pocket, and the cost
quickly becomes prohibitive.

This commit adds a special-case handling for PRs opened by
dependabot[bot].  For such PRs, we run a minimal set of CI checks,
mainly checking that the dependency updates do not break the build or
unit tests on Linux amd64, and that the code compiles on Mac and
Windows.

The full check will still run once the change lands in `master`, which
may turn out to also be a bit too expensive, but at least with this
change we cut that down a bit.