test: Add integration tests that actually drive the CLI and verify behavior (#34)
The initial tests are very basic, testing functionality that's already
reasonably well covered by the unit tests, but it lays the foundation
for much more ambitious testing.

At some point I want to have CI runners that exercise these tests in a
variety of Linux distros inside docker, thereby exploring compatibility
across a wide range of installations.

This also includes some unrelated docs changes, and some additional
lints were enabled. The config file command line args were revamped
considerably to now be more useful (especially in support of the
aforementioned integration tests that need to isolate each run in a temp
directory).

As part of the automated testing of the CLI, a bug in how args were
parsed was revealed, fixed, and tested. Now any arguments passed after
the crate name are assumed to be arguments to the crate and not to
`cgx`.