docs: Update README with installation instructions (#50)
My intention was to make specialized installation scripts as part of
this task, but it turns out that the ones generated by `cargo-dist` are
actually quite robust and probably sufficient for now.  So instead the
changes are to the README with revised installation instructions.

Whilst I had the hood up I also added more vanity shields to the README.

Closes #49