feat: Add `cargo-cgx` binary crate for cargo subcommand integration (#51)
This isn't, strictly speaking, necessary. The `cargo-cgx` binary has
literally the same `main()` function as the `cgx` binary, both deferring
to the `cgx` lib for the implementation.

Why duplicate the build effort and publish what is practically a
duplicate crate? A few reasons:

- Prevents typosquatting of the `cargo-cgx` name on crates.io. Part of
the integration of `cgx` into Rust tooling will be adding a `cargo cgx`
alias to the workspace, so users will habituate to `cargo cgx`. It's not
unreasonable to think that some might try to install `cargo-cgx`, and if
that happens it's important that they get the actual `cgx` and not some
malicious imposter.
- Perhaps some users will just prefer to use a `cargo` subcommand. Who
am I to judge?

Closes #14