refactor: DRY up the crate `Cargo.toml` files by using workspace package defaults
There was an excess of duplication in the `Cargo.toml` files for each crate in
the workspace. This commit moves common fields such as `edition`, `rust-version`,
`homepage`, `repository`, and `license` into the workspace `Cargo.toml` file
as workspace package defaults. Each crate's `Cargo.toml` file is updated to
reference these workspace defaults, reducing redundancy and simplifying future
maintenance.

In particular when we increment the MSRV we only have to do it in one
Cargo.toml file now instead of three.