refactor: Make our `insta` snapshot tests of SBOMs more robust
The tests were producing SBOMS that, correctly, included specific
versions of each transient dep.  The problem is that this is very
brittle.

This change takes advantage of `insta` functionality to post-process the
output to replace specific versions with placeholders.  This will still
break if we add new deps, but it should be stable as the existing deps
evolve.