ci: Update GHA runners to public free tiers
The last `dist` release build failed because it happened to coincide
with a ["brown-out" of the macos-13
runners](https://github.com/actions/runner-images/issues/13046).
Plus I recently made the repo public, meaning we can now use the free
runners for Mac and ARM platforms, which are much more cost-effective than
the custom runners I had to set up previously.

So it's time to change the runners that we use for CI builds and also
for `dist` release builds