ci: Config release-plz and dist to publish only binaries and in one release
The previous release, 0.0.5, was a bit of a shambles.  `release-plz`
made separate tags for each crate, and then `cargo-dist` made separate
releases for each crate.

That is wrong.  First of all, `cgx-core` is a library crate, there's
nothing to release but the source code archive.

Second, `cgx` and `cargo-cgx` should always be released together, as
they are versioned together and are in fact basically two names for the
same tool.

Hopefully this change fixes that, although testing it involves landing
the PR and finding out.