ci: Fix release-plz tagging
I incorrectly configured release-plz when I tried to prevent it from
creating separate tags per package.  I set all packages to make a
`v$VERSION` tag, causing all but the first one to fail to due that since
the tag already existed.  The correct way was a bit different.

This commit applies the approach documented in
https://release-plz.dev/docs/extra/single-tag