ci: Configure cargo-dist to exclude cargo-cgx from release text
I still want `cargo-cgx` binaries to be in the release, but I don't want
to see install instructions and changelog.  `cargo-cgx` is literally
just `cgx` but re-packaged as `cargo-cgx` for convenience for users who
prefer to run it that way.  It's distracting and confusing to have it
listed in the release text.