feat: Change how stderr output from `cargo build` is handled.
Before this feature branch was started, `stderr` output from `cargo
build` was just not visible to the user, so as a build was running there
was just a long awkward pause.

The initial implementation of the message reporting system changed that
to make stderr `inherited`, so it went directly to stderr of the `cgx`
process.  That was better, but it will break in the future when we
introduce progress reporting outputs.

This captures stderr output, and it as a message.  Then in `cgx_main`
it's handled.  When we add progress reporting in the future, `cgx_main`
can do so and decide how to properly render stderr output from `cargo
build`.