feat: Add structured message format for detailed operation reporting (#68)
This introduces a new `--message-format` CLI option that, if set to
`json`, causes cgx to emit structured JSON messages to stdout describing
its operations, including cache lookups, resolution results, downloads,
builds, and execution plans. This is useful for debugging, but it's
being added to allow more sophisticated integration tests that can
definitively assert whether or not the invoked `cgx` binary did what it
was supposed to do, be that hitting or not hitting the cache,
downloading or not downloading some specific artifacts, etc.

This is (for now) an unstable interface, and will not be considered when
making semver versioning decisions.

Closes #37