build: Change the `dist` profile to be more aggressive
This uses `lto=true` and `codegen-units = 1` to produce smaller and
faster binaries when building with the `dist` profile. The build will
take longer and have less concurrency, but in my tests it takes the
x86_64 linux binary from 21 to 15MB.  Even 15MB seems insanely large,
but it's an improvement.

I could strip debug symbols and configure abort on panic to shrink it
further but I'm not ready to take that step yet.  I still want
meaningful stack traces if something goes wrong.