feat: Add ability to resolve pre-built binaries for crates
This makes extensive changes to the config, cache, resolver, and
messsages modules to add the ability to resolve pre-built binaries for
crates, preferring those pre-built binaries where possible.

There are some complex rules around when a pre-built binary is
considered, and how pre-built binaries are discovered.  This isn't all
fully tested yet and is considered WIP.

Some definite todos:

- Scrape crates.io and make a list of the top N most popular binary
  crates and test this against them all; do we find pre-built binaries
  for every crate that publishes one?  If not why not?
- Extend the integration tests to use specially-selected crates to
  ensure that the github, gitlab, and quickinstall providers are all
  properly exercised.
- Add more dedicated tests exercising the various binary providers in
  isolation
- Consider adding retry logic to the binary providers, as they can get
  throttled or run into transient network issues (to say nothing of the
  very frequent github outages these days).

Somewhat unrelated to this main work, this commit also has some
improvements to the git client, which will now perform retries if it
gets throttled, and adds some integration tests specifically for that
pathway.  This isn't tied to the prebuilt binary functionality but it
came up as part of the testing I was doing on that feature so I include
it as a target of opportunity.

Closes #47